# apps/api/v1-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Firecrawl API",
    "version": "v1",
    "description": "API for interacting with Firecrawl services to perform web scraping and crawling tasks.",
    "contact": {
      "name": "Firecrawl Support",
      "url": "https://firecrawl.dev/support",
      "email": "support@firecrawl.dev"
    }
  },
  "servers": [
    {
      "url": "https://api.firecrawl.dev/v1"
    }
  ],
  "paths": {
    "/html-to-markdown": {
      "post": {
        "summary": "Convert HTML content to Markdown",
        "operationId": "convertHtmlToMarkdown",
        "tags": ["Conversion"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "html": {
                    "type": "string",
                    "description": "The HTML content to convert to Markdown"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "preserveHtmlTags": {
                        "type": "boolean",
                        "description": "Whether to preserve certain HTML tags in the output",
                        "default": false
                      }
                    }
                  }
                },
                "required": ["html"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful conversion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the conversion was successful"
                    },
                    "markdown": {
                      "type": "string",
                      "description": "The converted Markdown content"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    },
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/scrape": {
      "post": {
        "summary": "Scrape a single URL and optionally extract information using an LLM",
        "operationId": "scrapeAndExtractFromUrl",
        "tags": ["Scraping"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "The URL to scrape"
                  },
                  "formats": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "markdown",
                        "html",
                        "rawHtml",
                        "links",
                        "screenshot",
                        "extract",
                        "screenshot@fullPage"
                      ]
                    },
                    "description": "Formats to include in the output.",
                    "default": ["markdown"]
                  },
                  "onlyMainContent": {
                    "type": "boolean",
                    "description": "Only return the main content of the page excluding headers, navs, footers, etc.",
                    "default": true
                  },
                  "includeTags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Tags to include in the output."
                  },
                  "excludeTags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Tags to exclude from the output."
                  },
                  "headers": {
                    "type": "object",
                    "description": "Headers to send with the request. Can be used to send cookies, user-agent, etc."
                  },
                  "waitFor": {
                    "type": "integer",
                    "description": "Specify a delay in milliseconds before fetching the content, allowing the page sufficient time to load.",
                    "default": 0
                  },
                  "mobile": {
                    "type": "boolean",
                    "description": "Set to true if you want to emulate scraping from a mobile device. Useful for testing responsive pages and taking mobile screenshots.",
                    "default": false
                  },
                  "skipTlsVerification": {
                    "type": "boolean",
                    "description": "Skip TLS certificate verification when making requests",
                    "default": false
                  },
                  "timeout": {
                    "type": "integer",
                    "description": "Timeout in milliseconds for the request",
                    "default": 30000
                  },
                  "extract": {
                    "type": "object",
                    "description": "Extract object",
                    "properties": {
                      "schema": {
                        "type": "object",
                        "description": "The schema to use for the extraction (Optional)"
                      },
                      "systemPrompt": {
                        "type": "string",
                        "description": "The system prompt to use for the extraction (Optional)"
                      },
                      "prompt": {
                        "type": "string",
                        "description": "The prompt to use for the extraction without a schema (Optional)"
                      }
                    }
                  },
                  "actions": {
                    "type": "array",
                    "description": "Actions to perform on the page before grabbing the content",
                    "items": {
                      "oneOf": [
                        {
                          "type": "object",
                          "title": "Wait",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["wait"],
                              "description": "Wait for a specified amount of milliseconds"
                            },
                            "milliseconds": {
                              "type": "integer",
                              "minimum": 1,
                              "description": "Number of milliseconds to wait"
                            },
                            "selector": {
                              "type": "string",
                              "description": "Query selector to find the element by",
                              "example": "#my-element"
                            }
                          },
                          "required": ["type"]
                        },
                        {
                          "type": "object",
                          "title": "Screenshot",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["screenshot"],
                              "description": "Take a screenshot"
                            },
                            "fullPage": {
                              "type": "boolean",
                              "description": "Should the screenshot be full-page or viewport sized?",
                              "default": false
                            }
                          },
                          "required": ["type"]
                        },
                        {
                          "type": "object",
                          "title": "Click",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["click"],
                              "description": "Click on an element"
                            },
                            "selector": {
                              "type": "string",
                              "description": "Query selector to find the element by",
                              "example": "#load-more-button"
                            }
                          },
                          "required": ["type", "selector"]
                        },
                        {
                          "type": "object",
                          "title": "Write text",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["write"],
                              "description": "Write text into an input field, text area, or contenteditable element. Note: You must first focus the element using a 'click' action before writing. The text will be typed character by character to simulate keyboard input."
                            },
                            "text": {
                              "type": "string",
                              "description": "Text to type",
                              "example": "Hello, world!"
                            }
                          },
                          "required": ["type", "text"]
                        },
                        {
                          "type": "object",
                          "title": "Press a key",
                          "description": "Press a key on the page. See https://asawicki.info/nosense/doc/devices/keyboard/key_codes.html for key codes.",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["press"],
                              "description": "Press a key on the page"
                            },
                            "key": {
                              "type": "string",
                              "description": "Key to press",
                              "example": "Enter"
                            }
                          },
                          "required": ["type", "key"]
                        },
                        {
                          "type": "object",
                          "title": "Scroll",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["scroll"],
                              "description": "Scroll the page or a specific element"
                            },
                            "direction": {
                              "type": "string",
                              "enum": ["up", "down"],
                              "description": "Direction to scroll",
                              "default": "down"
                            },
                            "selector": {
                              "type": "string",
                              "description": "Query selector for the element to scroll",
                              "example": "#my-element"
                            }
                          },
                          "required": ["type"]
                        },
                        {
                          "type": "object",
                          "title": "Scrape",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["scrape"],
                              "description": "Scrape the current page content, returns the url and the html."
                            }
                          },
                          "required": ["type"]
                        },
                        {
                          "type": "object",
                          "title": "Execute JavaScript",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["executeJavascript"],
                              "description": "Execute JavaScript code on the page"
                            },
                            "script": {
                              "type": "string",
                              "description": "JavaScript code to execute",
                              "example": "document.querySelector('.button').click();"
                            }
                          },
                          "required": ["type", "script"]
                        }
                      ]
                    }
                  },
                  "location": {
                    "type": "object",
                    "description": "Location settings for the request. When specified, this will use an appropriate proxy if available and emulate the corresponding language and timezone settings. Defaults to 'US' if not specified.",
                    "properties": {
                      "country": {
                        "type": "string",
                        "description": "ISO 3166-1 alpha-2 country code (e.g., 'US', 'AU', 'DE', 'JP')",
                        "pattern": "^[A-Z]{2}$",
                        "default": "US"
                      },
                      "languages": {
                        "type": "array",
                        "description": "Preferred languages and locales for the request in order of priority. Defaults to the language of the specified location. See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Language",
                        "items": {
                          "type": "string",
                          "example": "en-US"
                        }
                      }
                    }
                  },
                  "removeBase64Images": {
                    "type": "boolean",
                    "description": "Removes all base 64 images from the output, which may be overwhelmingly long. The image's alt text remains in the output, but the URL is replaced with a placeholder."
                  }
                },
                "required": ["url"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Payment required to access this resource."
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Request rate limit exceeded. Please wait and try again later."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred on the server."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/batch/scrape/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "The ID of the batch scrape job",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Get the status of a batch scrape job",
        "operationId": "getBatchScrapeStatus",
        "tags": ["Scraping"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchScrapeStatusResponseObj"
                }
              }
            }
          },
          "402": {
            "description": "Payment required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Payment required to access this resource."
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Request rate limit exceeded. Please wait and try again later."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred on the server."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel a crawl job",
        "operationId": "cancelCrawl",
        "tags": ["Crawling"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Crawl job successfully cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["cancelled"],
                      "example": "cancelled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Crawl job not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Crawl job not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred on the server."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/extract": {
      "post": {
        "summary": "Extract structured data from pages using LLMs",
        "operationId": "extractData",
        "tags": ["Extraction"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "urls": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri",
                      "description": "The URLs to extract data from. URLs should be in glob format."
                    }
                  },
                  "prompt": {
                    "type": "string",
                    "description": "Prompt to guide the extraction process"
                  },
                  "schema": {
                    "type": "object",
                    "description": "Schema to define the structure of the extracted data",
                    "properties": {
                      "property1": {
                        "type": "string",
                        "description": "Description of property1"
                      },
                      "property2": {
                        "type": "integer",
                        "description": "Description of property2"
                      }
                    },
                    "required": ["property1", "property2"]
                  }
                },
                "required": ["urls", "prompt"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful extraction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid input data."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred on the server."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ExtractResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "<property1>": {
                "type": "string"
              },
              "<property2>": {
                "type": "number"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}